+++ src/components/themed/SearchFooter.tsx
@@ -55,16 +55,8 @@ export const SearchFooter = (props: SearchFooterProps) => {
     onChangeText(text)
   })
 
-  const handleSearchBlur = useHandler(() => {
-    if (searchText === '') {
-      onDoneSearching()
-    }
-  })
-
   const handleSearchClear = useHandler(() => {
-    if (!textInputRef.current?.isFocused()) {
-      onDoneSearching()
-    }
+    onDoneSearching()
   })
 
   const handleSearchFocus = useHandler(() => {
@@ -107,7 +99,7 @@ export const SearchFooter = (props: SearchFooterProps) => {
           placeholder={placeholder}
           onChangeText={handleSearchChangeText}
           value={searchText}
-          onBlur={handleSearchBlur}
+          active={isSearching}
           onClear={handleSearchClear}
           onFocus={handleSearchFocus}
           ref={textInputRef}